Track the awake status in esmetx
Previously we would honour an awake call regardless
of our current state, but this meant that we would
continously call backoff:deregister(), effectively
continously cancelling our timer and NEVER ACTUALLY
having the time to execute.

Now, we track our awake state, and if we're awake, we
just allow the current timer to actually run. If we're
not awake, we go ahead and call backoff:regular() which
resets things to normal and then set our new state as awake.

Signed-off-by: Essien Ita Essien <essiene@gmail.com>