Be forgiving of already started errors
Since each esmerx instance will call the Callback:init/0,
be forgiving of this error. The real solution is to
use a singleton starter centrally that the callback
does not have to be aware of.

also rename the file_appender. It seems all file_appenders
are in the global namespace and thus clash. Again log4erl
swallows errors without giving us the details.

Signed-off-by: Essien Ita Essien <essiene@gmail.com>