Do not use gen_server's implicit timeout
The semantics of the gen_server's implicit timeout during
init/1 is that it only occurs if NO MESSAGE is received in
the specified interval. In this case, the nanny is sending
messages to rxworker, so this is broken.

The proper thing is to send an explicit timeout message
after the required interval using erlang:send_after

Signed-off-by: Essien Ita Essien <essiene@gmail.com>