Define a #route_data{} record
This record will be used to return extra parsed params
from the routing table after the select route call.

The idea is that, during routing, the PDU will be traversed,
during this traversal, data will be saved into this record so
it can be grabbed back when needed later, without having to
traverse the PDU again, which is a potentially expensive process.

Signed-off-by: Essien Ita Essien <essiene@gmail.com>