Define the #soap_endpoint{} record
This record will hold the Soap Endpoing abstraction on the
serverside.

The idea is that when a new request is sent to the server for
dispatching, it will be sent to a particular end-point identified
by the name. The soap server should then fetch the #soap_endpoint{},
record and dispatch to the {M,F} resource in that end-point.

Signed-off-by: Essien Ita Essien <essiene@gmail.com>